% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdbconn.R
\name{dbd}
\alias{dbd}
\title{Disconnect from a PostgreSQL database}
\usage{
dbd(cn)
}
\arguments{
\item{cn}{A database connection object created by `dbc()`.}
}
\value{
Invisibly returns `NULL`.
}
\description{
Close a database connection created by `dbc()`.
}
