% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdbconn.R
\name{dir_rpg}
\alias{dir_rpg}
\title{Get the path to the rpgconn configuration directory}
\usage{
dir_rpg()
}
\value{
A character string giving the path to the configuration directory.
}
\description{
Returns the directory path used by rpgconn for storing configuration files.
}
