% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdbconn.R
\name{edit_config}
\alias{edit_config}
\title{Edit the connection configuration file}
\usage{
edit_config()
}
\value{
Invisibly returns the path to the configuration file.
}
\description{
In an interactive session, opens the YAML configuration file for editing using the user's
preferred editor. In non-interactive contexts (e.g. during automated checks), the file
path is returned invisibly without opening the editor.
}
