% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdbconn.R
\name{edit_options}
\alias{edit_options}
\title{Edit the options configuration file}
\usage{
edit_options()
}
\value{
Invisibly returns the path to the options file.
}
\description{
In an interactive session, opens the YAML options file for editing using the user's
preferred editor. In non-interactive contexts, the file path is returned invisibly
without opening the editor.
}
