% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdbconn.R
\name{init_yamls}
\alias{init_yamls}
\title{Initialize configuration files}
\usage{
init_yamls()
}
\value{
Invisibly returns the path to the configuration directory.
}
\description{
Ensures that the default connection (`config.yml`) and options (`options.yml`) templates
exist in the user-specific configuration directory. If they do not exist, they are copied
from the package's `extdata` directory. Use `edit_config()` and `edit_options()` to modify them.
}
