% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-config.R
\name{use_config}
\alias{use_config}
\title{Use New Config at Path}
\usage{
use_config(path, overwrite = FALSE)
}
\arguments{
\item{path}{path to new config yaml file to use}

\item{overwrite}{overwrite existing config yaml set prior. Default FALSE}
}
\value{
Invisible character vector giving the path to the active configuration
  file after it has been replaced.  In non-interactive sessions the file is
  silently returned.
}
\description{
Use New Config at Path
}
