% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-config.R
\name{use_config}
\alias{use_config}
\title{Use New Config at Path}
\usage{
use_config(path, overwrite = FALSE)
}
\arguments{
\item{path}{path to new config yaml file to use}

\item{overwrite}{overwrite existing config yaml set prior. Default FALSE}
}
\description{
Use New Config at Path
}
